adjusts

+++ src/controllers/MachineController/machine.controller.ts
@@ -14,10 +14,14 @@ import {
 import { Machine } from '@prisma/client';
 import { AuthenticatedRequest } from 'src/middlewares/authenticated.request';
 import { PrismaMachineRepository } from 'src/repositories/prisma/prisma.machine.repository';
+import { PrismaMonitoringPointsRepository } from 'src/repositories/prisma/prisma.monitoring.points';
 
 @Controller('machines')
 class MachineController {
-  constructor(private readonly machineRepository: PrismaMachineRepository) {}
+  constructor(
+    private readonly machineRepository: PrismaMachineRepository,
+    private readonly monitoringPointsRepository: PrismaMonitoringPointsRepository,
+  ) {}
 
   @Post()
   @HttpCode(HttpStatus.CREATED)
@@ -26,6 +30,11 @@ class MachineController {
     try {
       if (machine.type != 'FAN' && machine.type != 'PUMP') {
         throw new InternalServerErrorException('Invalid machine type. Verify.');
+      }
+      const existingMachine =
+        await this.machineRepository.getByMachine(machine);
+      if (existingMachine) {
+        throw new InternalServerErrorException('Machine already exists');
       }
       await this.machineRepository.create(machine);
       return {
@@ -67,7 +76,7 @@ class MachineController {
       } else {
         return {
           status: 'Ok!',
-          data: machine,
+          machine: machine,
         };
       }
     } catch (err) {
@@ -112,21 +121,31 @@ class MachineController {
   @HttpCode(HttpStatus.OK)
   @UseGuards(AuthenticatedRequest)
   async delete(@Param('id') id: number) {
-    console.log(id);
     try {
       const machine = await this.machineRepository.getById(Number(id));
+
       if (!machine) {
-        console.log(machine);
         throw new InternalServerErrorException('Machine not found');
       }
+
+      const hasMonitoringPoints =
+        await this.monitoringPointsRepository.getByMachine(Number(id));
+
+      if (hasMonitoringPoints) {
+        throw new InternalServerErrorException(
+          'Cannot delete machine; it is associated with monitoring points.',
+        );
+      }
+
       await this.machineRepository.delete(Number(id));
+
       return {
         status: 'Ok!',
         message: 'Machine deleted successfully',
       };
     } catch (err: any) {
       throw new InternalServerErrorException(
-        'Error deleting the machine, verify!. ' + err,
+        'Error deleting the machine, verify! ' + err,
       );
     }
   }
